requirements.txt
tika
translators
pyenchant
AnkiTools
pandas
openpyxl
genanki
gTTS
